Para desintaler Debian 11

+++ application/models/Participante_model.php
@@ -55,15 +55,52 @@ class Participante_model extends CI_model {
 
 
 
-  public function delete($id)
+  public function delete($idp,$ide)
 	{
- 		$this->db->where('idparticipante',$id);
-		$this->db->delete('participante');
-    		if($this->db->affected_rows()==1)
-			$result=true;
-		else
-			$result=false;
-		return $result;
+		$this->db->trans_start();
+		$condition = "idparticipante =" . $idp ;
+		$this->db->select('*');
+		$this->db->from('participante');
+		$this->db->where($condition);
+		$this->db->limit(1);
+		$query = $this->db->get();
+		if ($query->num_rows() != 0) {
+	 		$idpersona=$query->result()[0]->idpersona;
+		  $condition = "idpersona =" . $idpersona ;
+		  $this->db->select('*');
+		  $this->db->from('usuario');
+		  $this->db->where($condition);
+		  $this->db->limit(1);
+		  $query = $this->db->get();
+		  if ($query->num_rows() != 0) {
+	 		  $idusuario=$query->result()[0]->idusuario;
+ 		    $this->db->where('idusuario',$idusuario);
+ 		    $this->db->where('idevento',$ide);
+ 		    $this->db->where('onoff',1);
+		    $this->db->delete('password');
+
+	 		  $this->db->where('idparticipante',$idp);
+		    $this->db->delete('participante');
+        if($this->db->affected_rows()==1)
+        {
+            $this->db->trans_complete();
+			      $result=true;
+        }else{
+            $this->db->trans_complete();
+			      $result=false;
+        }
+     }else{	
+            $this->db->trans_complete();
+			      $result=false;
+     }
+   
+      }else{	
+
+            $this->db->trans_complete();
+			      $result=false;
+   }
+
+	return $result;
  	}
